Fix: missing tokens in MCP in memory transport
