+++ front/lib/actions/mcp_internal_actions/utils.ts
@@ -11,17 +11,12 @@ import type { MCPServerViewType } from "@app/lib/api/mcp";
 import type { OAuthProvider } from "@app/types";
 
 export function makeInternalMCPServer(
-  serverName: InternalMCPServerNameType,
-  options?: {
-    augmentedInstructions?: string;
-  }
+  name: InternalMCPServerNameType
 ): McpServer {
-  const serverInfo = getInternalMCPServerInfo(serverName);
-  const instructions =
-    options?.augmentedInstructions ?? serverInfo.instructions ?? undefined;
+  const serverInfo = getInternalMCPServerInfo(name);
 
   return new McpServer(serverInfo, {
-    instructions,
+    instructions: serverInfo.instructions ?? undefined,
   });
 }
