+++ front/lib/actions/mcp_metadata.ts
@@ -10,6 +10,11 @@ import type { JSONSchema7 as JSONSchema } from "json-schema";
 import type { ProxyAgent } from "undici";
 
 import {
+  DEFAULT_MCP_ACTION_DESCRIPTION,
+  DEFAULT_MCP_ACTION_NAME,
+  DEFAULT_MCP_ACTION_VERSION,
+} from "@app/lib/actions/constants";
+import {
   getConnectionForMCPServer,
   MCPServerPersonalAuthenticationRequiredError,
   MCPServerRequiresAdminAuthenticationError,
@@ -19,9 +24,13 @@ import {
   doesInternalMCPServerRequireBearerToken,
   getServerTypeAndIdFromSId,
 } from "@app/lib/actions/mcp_helper";
+import { DEFAULT_MCP_SERVER_ICON } from "@app/lib/actions/mcp_icons";
 import { connectToInternalMCPServer } from "@app/lib/actions/mcp_internal_actions";
+import {
+  getInternalMCPServerInfo,
+  getInternalMCPServerNameFromSId,
+} from "@app/lib/actions/mcp_internal_actions/constants";
 import { InMemoryWithAuthTransport } from "@app/lib/actions/mcp_internal_actions/in_memory_with_auth_transport";
-import { extractMetadataFromServerVersion } from "@app/lib/actions/mcp_metadata_extraction";
 import { MCPOAuthProvider } from "@app/lib/actions/mcp_oauth_provider";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import { ClientSideRedisMCPTransport } from "@app/lib/api/actions/mcp_client_side";
@@ -178,9 +187,15 @@ export async function connectToMCPServer(
                   })
                 : null;
 
-            const metadata = extractMetadataFromServerVersion(
-              mcpClient.getServerVersion()
+            const serverName = getInternalMCPServerNameFromSId(
+              params.mcpServerId
             );
+            if (!serverName) {
+              throw new Error(
+                `Internal server with id ${params.mcpServerId} do not resolve to a valid name.`
+              );
+            }
+            const serverInfo = getInternalMCPServerInfo(serverName);
 
             if (bearerTokenCredentials) {
               const authInfo: AuthInfo = {
@@ -197,7 +212,7 @@ export async function connectToMCPServer(
 
               client.setAuthInfo(authInfo);
               server.setAuthInfo(authInfo);
-            } else if (metadata.authorization) {
+            } else if (serverInfo.authorization) {
               if (!params.oAuthUseCase) {
                 throw new Error(
                   "Internal server requires authentication but no use case was provided - Should never happen"
@@ -243,8 +258,8 @@ export async function connectToMCPServer(
                 // Get conditional scope based on feature flag
                 const scope = await getConditionalScope(
                   auth,
-                  metadata.authorization.provider,
-                  metadata.authorization.scope
+                  serverInfo.authorization.provider,
+                  serverInfo.authorization.scope
                 );
 
                 if (params.oAuthUseCase === "personal_actions") {
@@ -264,7 +279,7 @@ export async function connectToMCPServer(
                     return new Err(
                       new MCPServerRequiresAdminAuthenticationError(
                         params.mcpServerId,
-                        metadata.authorization.provider,
+                        serverInfo.authorization.provider,
                         scope
                       )
                     );
@@ -272,7 +287,7 @@ export async function connectToMCPServer(
                   return new Err(
                     new MCPServerPersonalAuthenticationRequiredError(
                       params.mcpServerId,
-                      metadata.authorization.provider,
+                      serverInfo.authorization.provider,
                       scope
                     )
                   );
@@ -281,7 +296,7 @@ export async function connectToMCPServer(
                   return new Err(
                     new MCPServerRequiresAdminAuthenticationError(
                       params.mcpServerId,
-                      metadata.authorization.provider,
+                      serverInfo.authorization.provider,
                       scope
                     )
                   );
@@ -594,13 +609,17 @@ async function fetchRemoteServerMetaData(
 ): Promise<Result<Omit<MCPServerType, "sId">, Error>> {
   try {
     const serverVersion = mcpClient.getServerVersion();
-    const metadata = extractMetadataFromServerVersion(serverVersion);
 
     const toolsResult = await mcpClient.listTools();
     const serverTools = extractMetadataFromTools(toolsResult.tools);
 
     return new Ok({
-      ...metadata,
+      name: serverVersion?.name ?? DEFAULT_MCP_ACTION_NAME,
+      version: serverVersion?.version ?? DEFAULT_MCP_ACTION_VERSION,
+      description: serverVersion?.description ?? DEFAULT_MCP_ACTION_DESCRIPTION,
+      icon: DEFAULT_MCP_SERVER_ICON,
+      authorization: null,
+      documentationUrl: null,
       tools: serverTools,
       availability: "manual",
       allowMultipleInstances: true,
