+++ front/lib/api/mcp.ts
@@ -74,10 +74,13 @@ export type ClientSideMCPToolTypeWithStakeLevel =
   WithStakeLevelType<MCPToolWithAvailabilityType>;
 
 export type MCPServerType = {
-  sId: string;
+  // This will be part of the MCP server metadata at the protocol level.
   name: string;
   version: string;
   description: string;
+
+  // Everything below is only internal.
+  sId: string;
   icon: CustomResourceIconType | InternalAllowedIconType;
   authorization: AuthorizationInfo | null;
   tools: MCPToolType[];
